Refactor CSS and JavaScript code for adding copy functionality
This commit refactors the CSS and JavaScript code in the index.css and index.js files to add a new copy functionality. It introduces a new CSS class called "copy" in index.css and adds a new element with the class "copy" in index.js. The copy functionality allows users to copy the content of the textarea element to the clipboard. The commit also includes the addition of the copy-24.png icon in the data/icons directory.